app/requirements/base.txt
asgiref==3.9.1
channels==4.3.1
channels_redis==4.3.0
cloudinary==1.39.1
daphne==4.2.1
dj-database-url==2.1.0
django-cloudinary-storage==0.3.0
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.7
Faker==37.4.0
inflection==0.5.1
packaging==25.0
pillow==10.4.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
setuptools==80.9.0
sqlparse==0.5.3
toposort==1.10
typing_extensions==4.14.0
tzdata==2025.2
requests==2.32.4
uritemplate==4.2.0
